+++ Runtime/Scripts/Error.cs
@@ -5,15 +5,15 @@ namespace KindMen.Uxios
     public abstract class Error : Exception
     {
         public readonly Config Config;
-        public readonly Response Response;
+        public readonly IResponse Response;
 
-        protected Error(string message, Config config, Response response) : base(message)
+        protected Error(string message, Config config, IResponse response) : base(message)
         {
             Config = config;
             Response = response;
         }
 
-        protected Error(string message, Config config, Response response, Exception innerException) : base(message, innerException)
+        protected Error(string message, Config config, IResponse response, Exception innerException) : base(message, innerException)
         {
             Config = config;
             Response = response;