+++ Runtime/Scripts/Errors/DataProcessingError.cs
@@ -4,11 +4,11 @@ namespace KindMen.Uxios.Errors
 {
     public class DataProcessingError : Error
     {
-        public DataProcessingError(string message, Config config, Response response) : base(message, config, response)
+        public DataProcessingError(string message, Config config, IResponse response) : base(message, config, response)
         {
         }
 
-        public DataProcessingError(string message, Config config, Response response, Exception innerException) : base(message, config, response, innerException)
+        public DataProcessingError(string message, Config config, IResponse response, Exception innerException) : base(message, config, response, innerException)
         {
         }
     }