+++ Runtime/Scripts/Errors/Http/ForbiddenError.cs
@@ -4,7 +4,7 @@ namespace KindMen.Uxios.Errors.Http
 {
     public class ForbiddenError : AuthenticationError
     {
-        public ForbiddenError(string message, Config config, Response response, Exception e) : base(message, config, response, e)
+        public ForbiddenError(string message, Config config, IResponse response, Exception e) : base(message, config, response, e)
         {
         }
     }