+++ Runtime/Scripts/Errors/Http/UnauthorizedError.cs
@@ -4,7 +4,7 @@ namespace KindMen.Uxios.Errors.Http
 {
     public class UnauthorizedError : AuthenticationError
     {
-        public UnauthorizedError(string message, Config config, Response response, Exception e) : base(message, config, response, e)
+        public UnauthorizedError(string message, Config config, IResponse response, Exception e) : base(message, config, response, e)
         {
         }
     }