+++ Tests/Runtime/HttpAssertions.cs
@@ -5,7 +5,7 @@ namespace KindMen.Uxios.Tests
 {
     public static class HttpAssertions
     {
-        internal static void AssertStatusCode(Response response, HttpStatusCode httpStatusCode)
+        internal static void AssertStatusCode(IResponse response, HttpStatusCode httpStatusCode)
         {
             Assert.That(response.Status, Is.EqualTo(httpStatusCode));
         }